local/requirements.txt
azureml-core
azureml-dataprep
pandas